Added a language manager
+Added a LanguageManager class to store all strings for both english and italian languages.

Changed some small details about how the language is setted: it is now setted during the SplashScreen execution and is read directly by the static attribute of the statics class.

Adjusted PieChart size to better match with the screen, it is now clearer and bigger. Added some text views where there will be more specifics about the meal, like which food has the highest value for each nutritional values.